docs: generate llms.txt index for AI agents
Coding agents fetch /llms.txt to discover what documentation exists, but
mise.jdx.dev/llms.txt 404s today. Agents therefore work from training data
or guess, which is the root of most "the agent is bad at mise" reports.

Generate the index from the VitePress sidebar, with each description taken
from the lead paragraph of the page itself. No prose is written by hand and
no agent-specific documentation is introduced — it is a table of contents
pointing at the same pages a human reads, so it cannot drift from the docs.

Extract the sidebar into docs/.vitepress/sidebar.ts so the config and the
generator share one page list.
